smartcare-backend/requirements.txt
fastapi>=0.100.0
pydantic>=2.0.0
sqlalchemy>=1.4
asyncpg>=0.27
python-jose[cryptography]>=3.3
passlib[bcrypt]>=1.7
python-multipart>=0.0.6

# Real-time and AI client
redis>=5.0.0
httpx>=0.24.0

